starting code clean up
This project was just a place I slapped in a bunch of code I found and
code I wrote, along with some timing and validation code.
Now, I'm moving the individual kernels into their own testing functions
and subsections, to make it easier for others to find/use.
This check in is the basics - moved code out of main into functions,
created a test function at the top of the file that has the best/fastest
kernel that I wrote to solve the problem.  So it should be easier to
find it, rather than wading through the whole file.